Code refactor Update!
MapWidth and MapHeight are now int
Removed all program loop from game
Added Engine.cs, it controls program loop (render, update)
Player has static var LocalPlayer that points exaclty at player what we playing
New RegisterActor void in Engine.cs
Added IsInitialized to BaseObject
Added DestroyAll to Engine
Food are now is BaseObject, that means food is Destroyable
Player are now is also BaseObject
Text now have 999 Z-Index, that means Text must be on Top of render

+++ agar.io/Core/Types/GameConfiguration.cs
@@ -12,8 +12,8 @@ public static class GameConfiguration
 	public const int WindowWidth = 1280;
 	public const int WindowHeight = 720;
 
-	public const float MapWidth = 5000f;
-	public const float MapHeight = 5000f;
+	public const int MapWidth = 5000;
+	public const int MapHeight = 5000;
 
 	public const int MaxBots = 20;
 	public const int MaxFood = 800;
